test: adapt RPC configuration file test
The RPC configuration file test expects CRIU to error out if an unknown
option is found in the configuration file. As the previous commit of
this series changes unknown options in the configuration file (via RPC)
from an error to a warning, the corresponding test is now also changed
to not expect an error from CRIU but it is now looking for a warning in
the log file.

Signed-off-by: Adrian Reber <areber@redhat.com>