Fix RPC configuration file handling
While writing runc test cases to verify that runc correctly uses RPC
configuration files it became clear that some things were not working as
they are supposed to. Looking closer at the code to set log files
via RPC configuration files I discovered that the code seems wrong (at
least I did not understand it any more (or the intentions behind it)).

This code tries to simplify that logic a bit and add more comments to
make clear what the intentions of the RPC configuration file code is.

At the same time this changes/fixes the test. The relevant test case
test_rpc_with_configuration_file_overwriting_rpc() was actually designed
around the broken behaviour. It was only working if a previous
configuration file (set via environment variable in this case) and the
RPC configuration file have the same name. The test case which tests
that RPC configuration file settings are overwriting direct RPC settings
now makes sure that no other configuration file is set via the
environment variable. If it would be set, the test case would still
succeed, even with this patch applied. Which is and which was the
correct behaviour.

v2:
  - fix existing test case to test better (more correct)
  - make changes requested by Andrei

Signed-off-by: Adrian Reber <areber@redhat.com>