cr-restore: Mount proc only for NEWNS or NEWPID
In commit a201f59 (rst: Create custom proc every time we restore a
namespace) was introduced an approach to restore a mount namespace
using custom proc mount. This approach is also used to restore the
PID namespace.

However, there could be the case that neither of those namespaces is
being used. While commit 4dd0735 (restore: Save some time mounting
proc) handles the case when there are no namespaces to be restored,
this patch changes the CRIU's behaviour to use custom proc mount only
when it is needed (i.e. to restore mount or pid namespace).

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>