pstree: Restore task group leader when inherit SID
The current behaviour of CRIU is to inherit the group leader of the
parent process when migrating a session leader (with --shell-job).
However, it is possible for a process to be group leader without being
a session leader (e.g. a shell process). In this case CRIU should
restore the original group ID of the process and inherit only the
session ID.

Closes #593

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>