zdtm/thread-bomb: Limit stack size in thread-bomb
ia32 thread-bomb test failed when compel refused to seize the test,
trying to mmap() in remote process and getting ENOMEM.
It turns to be true - remote process thread-bomb was filled with 8Mb
mappings created by pthread_create() (the default stack size).

So, that 1024 * 8Mb is a bit too much to place in 4Gb.

Fix the test on 32-bit platforms by using much smaller stack.
Also check the return value of pthread_create().

Signed-off-by: Dmitry Safonov <dima@arista.com>